Split PullItem into View and structure for use in PullrequestItemFromRef
