+++ src/pullrequest/data/PullRequestFromRefQuery.js
@@ -1,3 +1,5 @@
+import PullRequestItemPartial from './PullRequestItemPartial';
+
 export default `
 query PullRequestFromRefQuery($owner: String!, $name: String!, $ref: String!) {
   repository(owner: $owner, name: $name) {
@@ -5,8 +7,7 @@ query PullRequestFromRefQuery($owner: String!, $name: String!, $ref: String!) {
       id
       associatedPullRequests(first:1) {
         nodes {
-          title
-          number
+            ${PullRequestItemPartial}
         }
       }
     }
